requirements.txt
absl-py==2.0.0
aiofiles==23.2.1
annotated-types==0.6.0
anyio==4.2.0
astunparse==1.6.3
bidict==0.22.1
cachetools==5.3.1
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.7
contourpy==1.1.1
cycler==0.12.0
d6tcollect==1.0.6
d6tflow==0.2.6
docutils==0.20.1
exceptiongroup==1.2.0
fastapi==0.108.0
fastapi-socketio==0.0.10
filelock==3.12.4
flatbuffers==23.5.26
fonttools==4.43.0
fsspec==2023.9.2
gast==0.5.4
google-api-core==2.11.1
google-api-python-client==2.85.0
google-api-python-client-py3==1.2
google-auth==2.23.0
google-auth-httplib2==0.1.1
google-auth-oauthlib==1.0.0
google-cloud-api-gateway==1.7.1
google-pasta==0.2.0
googleapis-common-protos==1.60.0
grpcio==1.58.0
grpcio-status==1.58.0
h11==0.14.0
h5py==3.9.0
httpcore==1.0.2
httplib2==0.22.0
httptools==0.6.1
httpx==0.26.0
huggingface-hub==0.17.3
idna==3.4
ifaddr==0.2.0
itsdangerous==2.1.2
jellyfish==1.0.1
Jinja2==3.1.2
joblib==1.3.2
keras==2.14.0
kiwisolver==1.4.5
libclang==16.0.6
lockfile==0.12.2
luigi==3.3.0
Markdown==3.4.4
markdown2==2.4.10
MarkupSafe==2.1.3
matplotlib==3.8.0
ml-dtypes==0.2.0
mpmath==1.3.0
networkx==3.1
nicegui==1.4.8
nltk==3.8.1
numpy==1.26.0
oauthlib==3.2.2
opt-einsum==3.3.0
orjson==3.9.10
packaging==23.1
pandas==2.1.1
Pillow==10.0.1
proto-plus==1.22.3
protobuf==4.24.3
pscript==0.7.7
pyarrow==13.0.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pydantic==2.5.3
pydantic_core==2.14.6
Pygments==2.17.2
pyparsing==3.1.1
python-daemon==3.0.1
python-dateutil==2.8.2
python-dotenv==1.0.0
python-engineio==4.8.1
python-multipart==0.0.6
python-socketio==5.10.0
pytube==15.0.0
pytz==2023.3.post1
PyYAML==6.0.1
regex==2023.8.8
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
safetensors==0.3.3
scikit-learn==1.3.1
scipy==1.11.2
scrapetube==2.5.1
seaborn==0.13.0
simple-websocket==1.0.0
six==1.16.0
sklearn==0.0.post9
sniffio==1.3.0
starlette==0.32.0.post1
sympy==1.12
tenacity==8.2.3
tensorboard==2.14.1
tensorboard-data-server==0.7.1
tensorflow==2.14.0
tensorflow-estimator==2.14.0
tensorflow-io-gcs-filesystem==0.34.0
termcolor==2.3.0
threadpoolctl==3.2.0
tokenizers==0.13.3
torch==2.0.1
tornado==6.3.3
tqdm==4.66.1
transformers==4.33.3
typing_extensions==4.8.0
tzdata==2023.3
uritemplate==4.1.1
urllib3==1.26.16
uvicorn==0.22.0
uvloop==0.19.0
vbuild==0.8.2
watchfiles==0.21.0
websockets==12.0
Werkzeug==2.3.7
wrapt==1.14.1
wsproto==1.2.0
youtube-transcript-api==0.6.0
